+++ lib_ft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: bvelonja <bvelonja@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/07 13:43:50 by bvelonja          #+#    #+#             */
-/*   Updated: 2025/03/14 16:10:29 by bvelonja         ###   ########.fr       */
+/*   Updated: 2025/03/26 01:34:25 by bvelonja         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char	*ft_strnstr(const char *str, const char *to_find, size_t len)
 		{
 			while (tmp[i] == to_find[i] && to_find[i])
 				i++;
-			if (to_find[i] == '\0' && len >= 0)
+			if (to_find[i] == '\0' && (int) len >= 0)
 				return ((char *)tmp);
 		}
 		tmp++;
